% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascendtSNE.R
\name{ascendtSNE.gene}
\alias{ascendtSNE.gene}
\title{Plotting and coloring ascend tSNE plots when samples/batch 1 and 2 are WT and TG respectively..}
\usage{
ascendtSNE.gene(gene, data = dat3d, exprs = exp.unlog.cln, cutOff = 0,
  alpha = 0.5, size = 1.5, option = "D", begin = 0.2, end = 0.5,
  direction = 1, lab_size = 4, heat = FALSE, manual_col = FALSE)
}
\arguments{
\item{gene}{Gene name.}

\item{data}{tSNE data table containing coordinate and cell information.}

\item{exprs}{Expression matrix with genes in rows and cells in columns.}

\item{cutOff}{Expression value cut off threshold. Default is "0".}

\item{alpha}{Transparency for geom_point. Default is "0.5".}

\item{option}{see ?viridis. Default is "D".}

\item{begin}{see ?viridis. Default is "0.2".}

\item{end}{see ?viridis. Default is "0.5".}

\item{direction}{see ?viridis. Default is "1".}

\item{lab_size}{Size of plot labels. Default is "4".}

\item{heat}{Plot as heatmap? TRUE/FALSE}

\item{manual_col}{Color option. Default is FALSE. Replace with string of colours or choose preset options. Options are: FALSE = viridis colour scale, TRUE = ggplot default palette, OrPU = Orange and purple, GrRd=Green and Red.}

\item{cell_size}{Size of geom_point. Default is "0.5".}

\item{show_branch_points}{See ?plot_cell_trajectory. Default is FALSE.}

\item{scale_to_expr}{TRUE/FALSE. True applies a log2 transformation to the counts. Default is FALSE.}
}
\value{
Plotting and coloring ascend tSNE plots.
}
\description{
Plotting and coloring ascend tSNE plots when samples/batch 1 and 2 are WT and TG respectively..
}
\examples{
em.set <- readRDS('em.set.RDS')
tsne.plot <- PlotTSNE(em.set, PCA = TRUE, condition = "cluster", seed = 122, perplexity = 30, theta = 0.5) 
# expression matrix
exp.unlog.cln <- as.matrix(GetExpressionMatrix(em.set))
# tSNE coordinates and cell information from PlotTSNE and em.set from ascend
dat3d <- cbind(tsne.plot$data,em.set@CellInformation)
colnames(dat3d) <- c('x','y','conditions','cell_barcode','batch','cluster')
ascendtSNE.gene('Cd207')            # plots which cells express > 0 transcript counts
ascendtSNE.gene('Cd207', heat=TRUE) # overlay heatmap of expression level over tSNE plot
ascendtSNE.info('cluster')          # plots cluster information
ascendtSNE.info('batch')            # plots batch/sample information
}
