% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon.R
\name{codon}
\alias{codon}
\title{converts codon nucleotide sequences to amino acids}
\usage{
codon(x, y = NULL)
}
\arguments{
\item{x}{first sequence}

\item{y}{default = NULL. If provided, will print aa1>aa2.}
}
\value{
converted amino acid annotation
}
\description{
converts codon nucleotide sequences to amino acids
}
\examples{
codon("ttt")
# [1] "Phe"
codon("ttt", "ttg")
# [1] "Phe>Leu"
}
