% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorMyTrajectory.R
\name{colorMyTrajectory}
\alias{colorMyTrajectory}
\title{Plotting and coloring Monocle 2 trajectory plot.}
\usage{
colorMyTrajectory(gene, data = HSMM, exprs = exp.unlog.cln,
  manual_col = FALSE, cutOff = 0, cell_size = 0.5,
  show_branch_points = FALSE, option = "D", begin = 0.2, end = 0.5,
  direction = 1, log2 = FALSE)
}
\arguments{
\item{gene}{Gene name.}

\item{data}{Monocle CellDataSet for the experiment.}

\item{exprs}{Expression matrix with genes in rows and cells in columns.}

\item{manual_col}{Color option. Default is FALSE. Replace with string of colours or choose preset options. Options are: FALSE = viridis colour scale, TRUE = ggplot default palette, OrPU = Orange and purple, GrRd=Green and Red.}

\item{cutOff}{Expression value cut off threshold. Default is "0".}

\item{cell_size}{Size of geom_point. Default is "0.5".}

\item{show_branch_points}{See ?plot_cell_trajectory. Default is FALSE.}

\item{option}{see ?viridis. Default is "D".}

\item{begin}{see ?viridis. Default is "0.2".}

\item{end}{see ?viridis. Default is "0.5".}

\item{direction}{see ?viridis. Default is "1".}

\item{scale_to_expr}{TRUE/FALSE. True applies a log2 transformation to the counts. Default is FALSE.}
}
\value{
Plotting and coloring Monocle 2 trajectory plot.
}
\description{
Plotting and coloring Monocle 2 trajectory plot.
}
\examples{
colorMyTrajectory('Il33', manual_col="OrPu")
colorMyTrajectory('Il33', manual_col="GrRd", show_branch_points=TRUE)
}
