% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDir copy.R, R/createDir.R
\name{createDir}
\alias{createDir}
\title{shortcut to create directory}
\usage{
createDir(directory)

createDir(directory)
}
\arguments{
\item{directory}{Directory to create}

\item{directory}{Directory to create}
}
\value{
creates directory if not found

creates directory if not found
}
\description{
shortcut to create directory

shortcut to create directory
}
\examples{
createDir('./plots/test/test2/test3/test4')
createDir('./plots/test/test2/test3/test4')
}
