% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirCreate.R
\name{dirCreate}
\alias{dirCreate}
\title{shortcut to create directory}
\usage{
dirCreate(directory)
}
\arguments{
\item{directory}{Directory to create}
}
\value{
creates directory if not found
}
\description{
shortcut to create directory
}
\examples{
dirCreate('./plots/test/test2/test3/test4')
}
