% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_hue.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Generate default ggplot colour palette depending on the number of colours required.}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{Number.}
}
\value{
Generate default ggplot colour palette depending on the number of colours required.
}
\description{
Generate default ggplot colour palette depending on the number of colours required.
}
\examples{
col <- gg_color_hue(5)
col
# [1] "#F8766D" "#A3A500" "#00BF7D" "#00B0F6" "#E76BF3"
}
