% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapk.R
\name{heatmap.k}
\alias{heatmap.k}
\title{Plotting a heatmap using pheatmap}
\usage{
heatmap.k(data, color = "RdWhBlu", scale = "row", show_colnames = TRUE,
  show_rownames = TRUE, cluster_rows = TRUE, cluster_cols = TRUE,
  fontsize = 9, n = 50, option = "A", direction = 1)
}
\arguments{
\item{data}{A data frame containing values of interest. Genes in rows, samples across columns.}

\item{color}{A string for what colours you want. e.g. c("blue","white,"red")}

\item{scale}{see ?pheatmap}

\item{show_colnames}{see ?pheatmap}

\item{show_rownames}{see ?pheatmap}

\item{cluster_rows}{see ?pheatmap}

\item{fontsize}{see ?pheatmap}

\item{n}{see ?viridis}

\item{option}{see ?viridis}

\item{direction}{see ?viridis}

\item{clsuter_cols}{see ?pheatmap}
}
\value{
A wrapper for pheatmap.
}
\description{
Plotting a heatmap using pheatmap
}
\examples{
data(iris)
data <- iris[,c(1:4)]
pheatmapping(data)
}
