% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_Kelvinny.R
\name{init_Kelvinny}
\alias{init_Kelvinny}
\title{documenting kelvinny shortcut}
\usage{
init_Kelvinny()
}
\value{
quickly redocumenting kelvinny
}
\description{
documenting kelvinny shortcut
}
\examples{
init_Kelvinny() # leaves folder, install, and change back to folder
}
