% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelvinny.R
\name{kelvinny}
\alias{kelvinny}
\title{shortcut to initiate building of kelvinny}
\usage{
kelvinny()
}
\value{
runs init kelvinny
}
\description{
shortcut to initiate building of kelvinny
}
\examples{
kelvinny()
}
