% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtx_to_h5.R
\name{mtx_to_h5}
\alias{mtx_to_h5}
\title{Saves matrix to .h5 format quickly}
\usage{
mtx_to_h5(object, filename, datasetname = "counts", ...)
}
\arguments{
\item{object}{input matrix (or maybe even data frame) - can be sparse or dense, but this is to sort the problem with dense matrices}

\item{filename}{name of the h5 object to save to}

\item{datasetname}{name of dataset}

\item{...}{pass to rhdf5 associated functions}
}
\value{
Save large (dense) matrix to .h5 format
}
\description{
Saves matrix to .h5 format quickly
}
\examples{
mtx_to_h5(matrixA, "matrix.h5")
mtx_to_h5totxt(matrixA, "matrix.h5")
}
