% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gmt.R
\name{parse_gmt}
\alias{parse_gmt}
\title{parse gmt file format as gmx}
\usage{
parse_gmt(file, header = TRUE, sep = "\\t", ...)
}
\arguments{
\item{file}{file}

\item{header}{default TRUE}

\item{sep}{default "\t"}

\item{...}{passes to read.csv}
}
\value{
reads a .gmt file like a .gmx file
}
\description{
parse gmt file format as gmx
}
