% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbcopy.R
\name{pbcopy}
\alias{pbcopy}
\title{a quick copy to/paste from clipboard. function}
\usage{
pbcopy(object)
}
\arguments{
\item{object}{anything object within R.}
}
\value{
Copy-pasta.
}
\description{
a quick copy to/paste from clipboard. function
}
