% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapk.R
\name{plotHeat}
\alias{plotHeat}
\title{Plotting a heatmap using pheatmap}
\usage{
plotHeat(dm, color = "RdWhBlu", scale = "row", show_colnames = TRUE,
  show_rownames = TRUE, cluster_rows = TRUE, cluster_cols = TRUE,
  fontsize = 9, n = 50, option = "D", direction = 1)
}
\arguments{
\item{color}{A string for what colours you want. e.g. c("blue","white,"red")}

\item{scale}{see ?pheatmap}

\item{show_colnames}{see ?pheatmap}

\item{show_rownames}{see ?pheatmap}

\item{cluster_rows}{see ?pheatmap}

\item{fontsize}{see ?pheatmap}

\item{n}{see ?viridis}

\item{option}{see ?viridis}

\item{direction}{see ?viridis}

\item{data}{A data frame containing values of interest. Genes in rows, samples across columns.}

\item{clsuter_cols}{see ?pheatmap}
}
\value{
A wrapper for pheatmap.
}
\description{
Plotting a heatmap using pheatmap
}
\examples{
data(iris)
head(iris)
#   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#1          5.1         3.5          1.4         0.2  setosa
#2          4.9         3.0          1.4         0.2  setosa
#3          4.7         3.2          1.3         0.2  setosa
#4          4.6         3.1          1.5         0.2  setosa
#5          5.0         3.6          1.4         0.2  setosa
#6          5.4         3.9          1.7         0.4  setosa
df <- iris[,c(1:4)]
plotHeat(df)
}
