% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predScSimilarity.R
\name{predScSimilarity}
\alias{predScSimilarity}
\title{Predicts expression data for class probability/similarity with a trained binomial/multinomial glmnet logistic regression model}
\usage{
predScSimilarity(model, test, standardize = TRUE, l.min = FALSE, ...)
}
\arguments{
\item{model}{trained model.}

\item{test}{Seurat object, SummarizedExperiment object or expression matrix for prediction}

\item{standardize}{a logical value specifying whether or not to standardize the test matrix}

\item{l.min}{logical. FALSE will use lambda.1se for prediction}

\item{...}{pass to predict function see ?predict.glmnet}
}
\value{
Generates prediction.
}
\description{
Predicts expression data for class probability/similarity with a trained binomial/multinomial glmnet logistic regression model
}
\examples{
pred <- predScSimilarity(model, test.sce)
}
