% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFclassifier.R
\name{RFclassifier}
\alias{RFclassifier}
\title{Random Forest Classifcation for single-cell expression data or just general matrix}
\usage{
RFclassifier(
  training,
  training.genes = NULL,
  training.classes = NULL,
  verbose = TRUE,
  probability = FALSE,
  ...
)
}
\arguments{
\item{training}{training expression matrix.}

\item{training.genes}{training genes to build classifier.}

\item{training.classes}{training classes to build classifier.}

\item{verbose}{verbosity.}

\item{probability}{Default = FALSE, which returns class when predicting. Otherwise TRUE returns probability estimates.}

\item{...}{pass to ranger::ranger.}
}
\value{
Training random forest classification model.
}
\description{
Random Forest Classifcation for single-cell expression data or just general matrix
}
\examples{
classifier <- RFclassifier(train.seurat, training.classes = train.seurat@ident, importance = "impurity")
prediction <- RFpredictor(classifier, test@data)
}
