% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFpredictor.R
\name{RFpredictor}
\alias{RFpredictor}
\title{Random Forest Classifcation for single-cell expression data or just general matrix}
\usage{
RFpredictor(classifier, test, ...)
}
\arguments{
\item{classifier}{trained RF classifier.}

\item{test}{test expression matrix.}

\item{...}{pass to predict.ranger.}
}
\value{
Predicting using random forest classification.
}
\description{
Random Forest Classifcation for single-cell expression data or just general matrix
}
\examples{
classifier <- RFclassifier(seurat, training.classes = seurat@ident, importance = "impurity")
prediction <- RFpredictor(classifier, test@data)
}
