% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_bootstrap.R
\name{similarity_bootstrap}
\alias{similarity_bootstrap}
\title{Bootstraps training and prediction of scRNA-seq data using binomial/multinomial logistic regression with tunable penalties}
\usage{
similarity_bootstrap(trainingSet, trainingCellType, testingSet,
  nboots = 50, simplify = FALSE, bs_nCores = 10,
  train_nCores = parallel::detectCores(), verbose = FALSE, ...)
}
\arguments{
\item{trainingSet}{Seurat object, SummarizedExperiment object or expression matrix for training}

\item{trainingCellType}{The cell types/clusters in the training data set}

\item{testingSet}{Seurat object, SummarizedExperiment object or expression matrix for testing}

\item{nboots}{integet specifying number of bootstrap runs to run}

\item{simplify}{logical. collapse the predictions to a single data frame}

\item{bs_nCores}{integer specifying number of cores for parallelization (for bootstrap).}

\item{train_nCores}{integer specifying number of cores for parallelization (for glmnet).}

\item{verbose}{logical. if TRUE, print the training/prediction steps. Otherwise, minimal messages will be returned on the screen}

\item{...}{other functions pass to glmnet and predict. see ?glmnet and ?predict.glmnet}
}
\value{
Generates prediction.
}
\description{
Bootstraps training and prediction of scRNA-seq data using binomial/multinomial logistic regression with tunable penalties
}
\examples{
pred <- predScSimilarity(model, test.sce)
}
