% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.tSNE.R
\name{simple.tSNE}
\alias{simple.tSNE}
\title{Plotting and coloring tSNE plot.}
\usage{
simple.tSNE(gene, data = dat3d, exprs = exp.unlog.cln, alpha = 0.5,
  size = 1.5, orange = FALSE)
}
\arguments{
\item{gene}{Gene name.}

\item{data}{tSNE data table containing coordinate and cell information.}

\item{exprs}{Expression matrix with genes in rows and cells in columns.}

\item{alpha}{Transparency for geom_point. Default is "0.5".}

\item{size}{Size of geom_point. Default is "0.5".}

\item{orange}{TRUE/FALSE. TRUE is orange/purple palette. FALSE is red/green palette}
}
\value{
Plotting and coloring tSNE plot.
}
\description{
Plotting and coloring tSNE plot.
}
\examples{
simple.tSNE('Il33', orange=TRUE) # plots which cells express > 0 transcript counts using orange and purple palette
}
