% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_model_glmnet.R
\name{test_model_glmnet}
\alias{test_model_glmnet}
\title{Predicting new data using trained glmnet model}
\usage{
test_model_glmnet(model, new_data, type = "response", ...)
}
\arguments{
\item{model}{trained model.}

\item{new_data}{new expression data.}

\item{type}{Type of prediction needed. see ?predict.glmnet}

\item{...}{pass to predict. see ?predict.glmnet}
}
\value{
Generates prediction.
}
\description{
Predicting new data using trained glmnet model
}
\examples{
pred <- test_model_glmnet(model = model, new_data = newdat, type = "link")
}
