% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model_glmnet.R
\name{train_model_glmnet}
\alias{train_model_glmnet}
\title{Training a glmnet model for binomial prediction.}
\usage{
train_model_glmnet(
  data,
  variable_colname,
  alpha = 1,
  type.measure = "mse",
  type = "response",
  standardize = FALSE,
  cutOff = 0.5,
  nfolds = 10,
  ...
)
}
\arguments{
\item{data}{expression data with genes in rows and samples in columns.}

\item{variable_colname}{Column name for binary class.}

\item{alpha}{Default = 1 for Lasso penalty. 0 = Ridge penalty. between 0 - 1 = Elastic net. See ?cv.glmnet.}

\item{type.measure}{See ?cv.glmnet.}

\item{type}{term for internal prediction type.}

\item{standardize}{Default = FALSE. Expects data to be pre-scaled. if not, glmnet will do it for you.}

\item{cutOff}{value for internal prediction.}

\item{nfolds}{number of folds for cross validation. use ncol(data) for LOOCV.}

\item{...}{pass to glmnet}
}
\value{
Generates a glmnet model.
}
\description{
Training a glmnet model for binomial prediction.
}
\examples{
model <- train_model_glmnet(data, variable_colname = "disease", alpha = 0.5, cutOff = 0.5, nfolds = ncol(data)) # LOOCV Elastic net regularization
}
