% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeat.R
\name{plotHeat}
\alias{plotHeat}
\title{Plotting a heatmap using pheatmap}
\usage{
plotHeat(
  d,
  color = "BluWhRd",
  scale = "row",
  n = 50,
  alpha = 1,
  begin = 0,
  end = 1,
  option = "D",
  direction = 1,
  show_colnames = TRUE,
  show_rownames = TRUE,
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  drop_levels = TRUE,
  fontsize = 9,
  legend = TRUE,
  border_color = NA,
  ...
)
}
\arguments{
\item{color}{Default has "BluWhRd", "RdWhBlu", "viridis" or a string of the colours you want. e.g. c("blue","yellow","orange").}

\item{scale}{Default is "row". Other options include "column" or "none". see ?pheatmap.}

\item{n}{see ?viridis}

\item{alpha}{see?viridis}

\item{begin}{see ?viridis}

\item{end}{see ?viridis}

\item{option}{see ?viridis}

\item{direction}{see ?viridis}

\item{show_colnames}{TRUE/FALSE. see ?pheatmap.}

\item{show_rownames}{TRUE/FALSE. see ?pheatmap.}

\item{cluster_rows}{TRUE/FALSE. see ?pheatmap.}

\item{cluster_cols}{TRUE/FALSE. see ?pheatmap.}

\item{drop_levels}{TRUE/FALSE. see ?pheatmap.}

\item{fontsize}{Default is 9. see ?pheatmap.}

\item{legend}{TRUE/FALSE. see ?pheatmap.}

\item{border_color}{see ?pheatmap.}

\item{...}{Passes on arguments to pheatmap.}

\item{data}{A data frame containing values of interest. Genes in rows, samples across columns. if supplied as a matrix with rownames as genes and the 1st column is exprs values, it will still go through.}
}
\value{
A wrapper for pheatmap.
}
\description{
Plotting a heatmap using pheatmap
}
\examples{
data(iris)
head(iris)
#   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#1          5.1         3.5          1.4         0.2  setosa
#2          4.9         3.0          1.4         0.2  setosa
#3          4.7         3.2          1.3         0.2  setosa
#4          4.6         3.1          1.5         0.2  setosa
#5          5.0         3.6          1.4         0.2  setosa
#6          5.4         3.9          1.7         0.4  setosa
df <- iris[,c(1:4)]
plotHeat(df)

# using example from pheatmap
# Create test matrix
test = matrix(rnorm(200), 20, 10)
test[1:10, seq(1, 10, 2)] = test[1:10, seq(1, 10, 2)] + 3
test[11:20, seq(2, 10, 2)] = test[11:20, seq(2, 10, 2)] + 2
test[15:20, seq(2, 10, 2)] = test[15:20, seq(2, 10, 2)] + 4
colnames(test) = paste("Test", 1:10, sep = "")
rownames(test) = paste("Gene", 1:20, sep = "")
# Generate annotations for rows and columns
annotation_col = data.frame(
  CellType = factor(rep(c("CT1", "CT2"), 5)),
  Time = 1:5
)
rownames(annotation_col) = paste("Test", 1:10, sep = "")

annotation_row = data.frame(
  GeneClass = factor(rep(c("Path1", "Path2", "Path3"), c(10, 4, 6)))
)
rownames(annotation_row) = paste("Gene", 1:20, sep = "")
#plot!
plotHeat(test, annotation_col=annotation_col, annotation_row=annotation_row)
}
