% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predScSimilarity.R
\name{predScSimilarity}
\alias{predScSimilarity}
\title{Predicts scRNA-seq data from trained multinomial cv.glmnet model}
\usage{
predScSimilarity(model, test, standardize = TRUE, ...)
}
\arguments{
\item{model}{trained model.}

\item{test}{SingleCellExperiment object for prediction}

\item{standardize}{a logical value specifying whether or not to standardize the test matrix}

\item{...}{pass to predict. see ?predict.glmnet}
}
\value{
Generates prediction.
}
\description{
Predicts scRNA-seq data from trained multinomial cv.glmnet model
}
\examples{
pred <- predScSimilarity(model, test.sce)
}
