% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainScSimilarity.R
\name{trainScSimilarity}
\alias{trainScSimilarity}
\title{Trains scRNA-seq data to be used as multinomial glmnet::cv.glmnet model}
\usage{
trainScSimilarity(train_data, train_cell_type, train_genes = NULL,
  standardize = TRUE, nfolds = 10, alpha = 0.99,
  nParallel = parallel::detectCores(), ...)
}
\arguments{
\item{train_data}{seurat object, SummarizedExperiment object or expression matrix for training}

\item{train_cell_type}{The cell types/clusters in the training data set}

\item{train_genes}{Genes to use for training. If not provided, it will try to pick from all genes in the training dataset as per default glmnet.}

\item{standardize}{a logical value specifying whether or not to standardize the train matrix}

\item{nfolds}{integer specifying bin for cross validation. Use all samples if doing LOOCV.}

\item{alpha}{regularization parameter}

\item{nParallel}{integer specifying number of cores for parallelization.}

\item{...}{pass to glmnet}
}
\value{
Generates a trained model for predicting cell types for scRNAseq data
}
\description{
Trains scRNA-seq data to be used as multinomial glmnet::cv.glmnet model
}
\examples{
fit <- trainScSimilarity(data, cluster, genes)
}
